test: organise backages
