+++ src/test/kotlin/logic/useCase/UpdateProjectUseCaseTest.kt
@@ -1,4 +1,4 @@
-package logic.usecases
+package logic.useCase
 
 import com.google.common.truth.Truth.assertThat
 import io.mockk.every
@@ -7,7 +7,7 @@ import org.example.logic.models.*
 import org.example.logic.repositries.AuditLogRepository
 import org.example.logic.repositries.AuthenticationRepository
 import org.example.logic.repositries.ProjectRepository
-import org.example.logic.usecases.UpdateProjectUseCase
+import org.example.logic.useCase.UpdateProjectUseCase
 import org.example.logic.utils.*
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
